+++ EDDIVAPlugin/VoiceAttackPlugin.cs
@@ -402,6 +402,17 @@ namespace EDDIVAPlugin
                     setDecimal(ref decimalValues, "Ship health", Cmdr.Ship.Health);
                     setInt(ref intValues, "Ship cargo capacity", Cmdr.Ship.CargoCapacity);
                     setInt(ref intValues, "Ship cargo carried", Cmdr.Ship.CargoCarried);
+                    // Add number of limpets carried
+                    int limpets = 0;
+                    foreach (Cargo cargo in Cmdr.Ship.Cargo)
+                    {
+                        if (cargo.Commodity.Name == "Limpet")
+                        {
+                            limpets += cargo.Quantity;
+                        }
+                    }
+                    setInt(ref intValues, "Ship limpets carried", limpets);
+
 
                     SetModuleDetails("Ship bulkheads", Cmdr.Ship.Bulkheads, ref textValues, ref intValues, ref decimalValues);
                     SetOutfittingCost("Ship bulkheads", Cmdr.Ship.Bulkheads, ref Cmdr.Outfitting, ref textValues, ref decimalValues);