+++ ChangeLog.md
@@ -1,5 +1,6 @@
   * Added configuration option to set insurance excess percentage
   * Added Insurance (decimal): the percentage insurance excess (usually 5, 3.75 or 2.5)
+  * Added Ship limpets carried (int): the number of limpets carried by the ship
 
 #1.0.0
   * Fix minor VoiceAttack script issues